+++ src/OpenTelemetry.AutoInstrumentation/AutoInstrumentationEventSource.cs
@@ -21,26 +21,43 @@ namespace OpenTelemetry.AutoInstrumentation;
 /// <summary>
 /// EventSource implementation for OpenTelemetry AutoInstrumentation implementation.
 /// </summary>
-[EventSource(Name = "OpenTelemetry-AutoInstrumentation")]
-internal class AutoInstrumentationEventSource : EventSource
+internal partial class AutoInstrumentationEventSource : EventSource
 {
-#pragma warning disable SA1401 // Fields should be private
-    public static readonly AutoInstrumentationEventSource Log = new AutoInstrumentationEventSource();
-#pragma warning restore SA1401 // Fields should be private
-
-    /// <summary>Logs as Trace level message.</summary>
-    /// <param name="message">Message to log.</param>
-    [Event(1, Message = "{0}", Level = EventLevel.Informational)]
-    public void Trace(string message)
+    private AutoInstrumentationEventSource()
     {
-        this.WriteEvent(1, message);
     }
+
+    public static AutoInstrumentationEventSource Log { get; } = new();
 
     /// <summary>Logs as Error level message.</summary>
     /// <param name="message">Error to log.</param>
     [Event(2, Message = "{0}", Level = EventLevel.Error)]
     public void Error(string message)
     {
-        this.WriteEvent(2, message);
+        WriteEvent(2, message);
+    }
+
+    /// <summary>Logs as Warning level message.</summary>
+    /// <param name="message">Message to log.</param>
+    [Event(3, Message = "{0}", Level = EventLevel.Warning)]
+    public void Warning(string message)
+    {
+        WriteEvent(3, message);
+    }
+
+    /// <summary>Logs as Information level message.</summary>
+    /// <param name="message">Message to log.</param>
+    [Event(4, Message = "{0}", Level = EventLevel.Informational)]
+    public void Information(string message)
+    {
+        WriteEvent(4, message);
+    }
+
+    /// <summary>Logs as Warning level message.</summary>
+    /// <param name="message">Message to log.</param>
+    [Event(5, Message = "{0}", Level = EventLevel.Verbose)]
+    public void Verbose(string message)
+    {
+        WriteEvent(5, message);
     }
 }