+++ src/OpenTelemetry.AutoInstrumentation/Logging/CustomLogger.cs
@@ -33,174 +33,195 @@ internal class CustomLogger : IOtelLogger
         return true;
     }
 
-    public void Debug(string messageTemplate)
-        => Write(LogLevel.Debug, exception: null, messageTemplate, NoPropertyValues);
+    public void Debug(string messageTemplate, bool writeToEventLog)
+        => Write(LogLevel.Debug, exception: null, messageTemplate, NoPropertyValues, writeToEventLog);
 
-    public void Debug<T>(string messageTemplate, T property)
-        => Write(LogLevel.Debug, exception: null, messageTemplate, property);
+    public void Debug<T>(string messageTemplate, T property, bool writeToEventLog)
+        => Write(LogLevel.Debug, exception: null, messageTemplate, property, writeToEventLog);
 
-    public void Debug<T0, T1>(string messageTemplate, T0 property0, T1 property1)
-        => Write(LogLevel.Debug, exception: null, messageTemplate, property0, property1);
+    public void Debug<T0, T1>(string messageTemplate, T0 property0, T1 property1, bool writeToEventLog)
+        => Write(LogLevel.Debug, exception: null, messageTemplate, property0, property1, writeToEventLog);
 
-    public void Debug<T0, T1, T2>(string messageTemplate, T0 property0, T1 property1, T2 property2)
-        => Write(LogLevel.Debug, exception: null, messageTemplate, property0, property1, property2);
+    public void Debug<T0, T1, T2>(string messageTemplate, T0 property0, T1 property1, T2 property2, bool writeToEventLog)
+        => Write(LogLevel.Debug, exception: null, messageTemplate, property0, property1, property2, writeToEventLog);
 
-    public void Debug(string messageTemplate, object[] args)
-        => Write(LogLevel.Debug, exception: null, messageTemplate, args);
+    public void Debug(string messageTemplate, object[] args, bool writeToEventLog)
+        => Write(LogLevel.Debug, exception: null, messageTemplate, args, writeToEventLog);
 
-    public void Debug(Exception exception, string messageTemplate)
-        => Write(LogLevel.Debug, exception, messageTemplate, NoPropertyValues);
+    public void Debug(Exception exception, string messageTemplate, bool writeToEventLog)
+        => Write(LogLevel.Debug, exception, messageTemplate, NoPropertyValues, writeToEventLog);
 
-    public void Debug<T>(Exception exception, string messageTemplate, T property)
-        => Write(LogLevel.Debug, exception, messageTemplate, property);
+    public void Debug<T>(Exception exception, string messageTemplate, T property, bool writeToEventLog)
+        => Write(LogLevel.Debug, exception, messageTemplate, property, writeToEventLog);
 
-    public void Debug<T0, T1>(Exception exception, string messageTemplate, T0 property0, T1 property1)
-        => Write(LogLevel.Debug, exception, messageTemplate, property0, property1);
+    public void Debug<T0, T1>(Exception exception, string messageTemplate, T0 property0, T1 property1, bool writeToEventLog)
+        => Write(LogLevel.Debug, exception, messageTemplate, property0, property1, writeToEventLog);
 
-    public void Debug<T0, T1, T2>(Exception exception, string messageTemplate, T0 property0, T1 property1, T2 property2)
-        => Write(LogLevel.Debug, exception, messageTemplate, property0, property1, property2);
+    public void Debug<T0, T1, T2>(Exception exception, string messageTemplate, T0 property0, T1 property1, T2 property2, bool writeToEventLog)
+        => Write(LogLevel.Debug, exception, messageTemplate, property0, property1, property2, writeToEventLog);
 
-    public void Debug(Exception exception, string messageTemplate, object[] args)
-        => Write(LogLevel.Debug, exception, messageTemplate, args);
+    public void Debug(Exception exception, string messageTemplate, object[] args, bool writeToEventLog)
+        => Write(LogLevel.Debug, exception, messageTemplate, args, writeToEventLog);
 
-    public void Information(string messageTemplate)
-        => Write(LogLevel.Information, exception: null, messageTemplate, NoPropertyValues);
+    public void Information(string messageTemplate, bool writeToEventLog)
+        => Write(LogLevel.Information, exception: null, messageTemplate, NoPropertyValues, writeToEventLog);
 
-    public void Information<T>(string messageTemplate, T property)
-        => Write(LogLevel.Information, exception: null, messageTemplate, property);
+    public void Information<T>(string messageTemplate, T property, bool writeToEventLog)
+        => Write(LogLevel.Information, exception: null, messageTemplate, property, writeToEventLog);
 
-    public void Information<T0, T1>(string messageTemplate, T0 property0, T1 property1)
-        => Write(LogLevel.Information, exception: null, messageTemplate, property0, property1);
+    public void Information<T0, T1>(string messageTemplate, T0 property0, T1 property1, bool writeToEventLog)
+        => Write(LogLevel.Information, exception: null, messageTemplate, property0, property1, writeToEventLog);
 
-    public void Information<T0, T1, T2>(string messageTemplate, T0 property0, T1 property1, T2 property2)
-        => Write(LogLevel.Information, exception: null, messageTemplate, property0, property1, property2);
+    public void Information<T0, T1, T2>(string messageTemplate, T0 property0, T1 property1, T2 property2, bool writeToEventLog)
+        => Write(LogLevel.Information, exception: null, messageTemplate, property0, property1, property2, writeToEventLog);
 
-    public void Information(string messageTemplate, object[] args)
-        => Write(LogLevel.Information, exception: null, messageTemplate, args);
+    public void Information(string messageTemplate, object[] args, bool writeToEventLog)
+        => Write(LogLevel.Information, exception: null, messageTemplate, args, writeToEventLog);
 
-    public void Information(Exception exception, string messageTemplate)
-        => Write(LogLevel.Information, exception, messageTemplate, NoPropertyValues);
+    public void Information(Exception exception, string messageTemplate, bool writeToEventLog)
+        => Write(LogLevel.Information, exception, messageTemplate, NoPropertyValues, writeToEventLog);
 
-    public void Information<T>(Exception exception, string messageTemplate, T property)
-        => Write(LogLevel.Information, exception, messageTemplate, property);
+    public void Information<T>(Exception exception, string messageTemplate, T property, bool writeToEventLog)
+        => Write(LogLevel.Information, exception, messageTemplate, property, writeToEventLog);
 
-    public void Information<T0, T1>(Exception exception, string messageTemplate, T0 property0, T1 property1)
-        => Write(LogLevel.Information, exception, messageTemplate, property0, property1);
+    public void Information<T0, T1>(Exception exception, string messageTemplate, T0 property0, T1 property1, bool writeToEventLog)
+        => Write(LogLevel.Information, exception, messageTemplate, property0, property1, writeToEventLog);
 
-    public void Information<T0, T1, T2>(Exception exception, string messageTemplate, T0 property0, T1 property1, T2 property2)
-        => Write(LogLevel.Information, exception, messageTemplate, property0, property1, property2);
+    public void Information<T0, T1, T2>(Exception exception, string messageTemplate, T0 property0, T1 property1, T2 property2, bool writeToEventLog)
+        => Write(LogLevel.Information, exception, messageTemplate, property0, property1, property2, writeToEventLog);
 
-    public void Information(Exception exception, string messageTemplate, object[] args)
-        => Write(LogLevel.Information, exception, messageTemplate, args);
+    public void Information(Exception exception, string messageTemplate, object[] args, bool writeToEventLog)
+        => Write(LogLevel.Information, exception, messageTemplate, args, writeToEventLog);
 
-    public void Warning(string messageTemplate)
-        => Write(LogLevel.Warning, exception: null, messageTemplate, NoPropertyValues);
+    public void Warning(string messageTemplate, bool writeToEventLog)
+        => Write(LogLevel.Warning, exception: null, messageTemplate, NoPropertyValues, writeToEventLog);
 
-    public void Warning<T>(string messageTemplate, T property)
-        => Write(LogLevel.Warning, exception: null, messageTemplate, property);
+    public void Warning<T>(string messageTemplate, T property, bool writeToEventLog)
+        => Write(LogLevel.Warning, exception: null, messageTemplate, property, writeToEventLog);
 
-    public void Warning<T0, T1>(string messageTemplate, T0 property0, T1 property1)
-        => Write(LogLevel.Warning, exception: null, messageTemplate, property0, property1);
+    public void Warning<T0, T1>(string messageTemplate, T0 property0, T1 property1, bool writeToEventLog)
+        => Write(LogLevel.Warning, exception: null, messageTemplate, property0, property1, writeToEventLog);
 
-    public void Warning<T0, T1, T2>(string messageTemplate, T0 property0, T1 property1, T2 property2)
-        => Write(LogLevel.Warning, exception: null, messageTemplate, property0, property1, property2);
+    public void Warning<T0, T1, T2>(string messageTemplate, T0 property0, T1 property1, T2 property2, bool writeToEventLog)
+        => Write(LogLevel.Warning, exception: null, messageTemplate, property0, property1, property2, writeToEventLog);
 
-    public void Warning(string messageTemplate, object[] args)
-        => Write(LogLevel.Warning, exception: null, messageTemplate, args);
+    public void Warning(string messageTemplate, object[] args, bool writeToEventLog)
+        => Write(LogLevel.Warning, exception: null, messageTemplate, args, writeToEventLog);
 
-    public void Warning(Exception exception, string messageTemplate)
-        => Write(LogLevel.Warning, exception, messageTemplate, NoPropertyValues);
+    public void Warning(Exception exception, string messageTemplate, bool writeToEventLog)
+        => Write(LogLevel.Warning, exception, messageTemplate, NoPropertyValues, writeToEventLog);
 
-    public void Warning<T>(Exception exception, string messageTemplate, T property)
-        => Write(LogLevel.Warning, exception, messageTemplate, property);
+    public void Warning<T>(Exception exception, string messageTemplate, T property, bool writeToEventLog)
+        => Write(LogLevel.Warning, exception, messageTemplate, property, writeToEventLog);
 
-    public void Warning<T0, T1>(Exception exception, string messageTemplate, T0 property0, T1 property1)
-        => Write(LogLevel.Warning, exception, messageTemplate, property0, property1);
+    public void Warning<T0, T1>(Exception exception, string messageTemplate, T0 property0, T1 property1, bool writeToEventLog)
+        => Write(LogLevel.Warning, exception, messageTemplate, property0, property1, writeToEventLog);
 
-    public void Warning<T0, T1, T2>(Exception exception, string messageTemplate, T0 property0, T1 property1, T2 property2)
-        => Write(LogLevel.Warning, exception, messageTemplate, property0, property1, property2);
+    public void Warning<T0, T1, T2>(Exception exception, string messageTemplate, T0 property0, T1 property1, T2 property2, bool writeToEventLog)
+        => Write(LogLevel.Warning, exception, messageTemplate, property0, property1, property2, writeToEventLog);
 
-    public void Warning(Exception exception, string messageTemplate, object[] args)
-        => Write(LogLevel.Warning, exception, messageTemplate, args);
+    public void Warning(Exception exception, string messageTemplate, object[] args, bool writeToEventLog)
+        => Write(LogLevel.Warning, exception, messageTemplate, args, writeToEventLog);
 
-    public void Error(string messageTemplate)
-        => Write(LogLevel.Error, exception: null, messageTemplate, NoPropertyValues);
+    public void Error(string messageTemplate, bool writeToEventLog)
+        => Write(LogLevel.Error, exception: null, messageTemplate, NoPropertyValues, writeToEventLog);
 
-    public void Error<T>(string messageTemplate, T property)
-        => Write(LogLevel.Error, exception: null, messageTemplate, property);
+    public void Error<T>(string messageTemplate, T property, bool writeToEventLog)
+        => Write(LogLevel.Error, exception: null, messageTemplate, property, writeToEventLog);
 
-    public void Error<T0, T1>(string messageTemplate, T0 property0, T1 property1)
-        => Write(LogLevel.Error, exception: null, messageTemplate, property0, property1);
+    public void Error<T0, T1>(string messageTemplate, T0 property0, T1 property1, bool writeToEventLog)
+        => Write(LogLevel.Error, exception: null, messageTemplate, property0, property1, writeToEventLog);
 
-    public void Error<T0, T1, T2>(string messageTemplate, T0 property0, T1 property1, T2 property2)
-        => Write(LogLevel.Error, exception: null, messageTemplate, property0, property1, property2);
+    public void Error<T0, T1, T2>(string messageTemplate, T0 property0, T1 property1, T2 property2, bool writeToEventLog)
+        => Write(LogLevel.Error, exception: null, messageTemplate, property0, property1, property2, writeToEventLog);
 
-    public void Error(string messageTemplate, object[] args)
-        => Write(LogLevel.Error, exception: null, messageTemplate, args);
+    public void Error(string messageTemplate, object[] args, bool writeToEventLog)
+        => Write(LogLevel.Error, exception: null, messageTemplate, args, writeToEventLog);
 
-    public void Error(Exception exception, string messageTemplate)
-        => Write(LogLevel.Error, exception, messageTemplate, NoPropertyValues);
+    public void Error(Exception exception, string messageTemplate, bool writeToEventLog)
+        => Write(LogLevel.Error, exception, messageTemplate, NoPropertyValues, writeToEventLog);
 
-    public void Error<T>(Exception exception, string messageTemplate, T property)
-        => Write(LogLevel.Error, exception, messageTemplate, property);
+    public void Error<T>(Exception exception, string messageTemplate, T property, bool writeToEventLog)
+        => Write(LogLevel.Error, exception, messageTemplate, property, writeToEventLog);
 
-    public void Error<T0, T1>(Exception exception, string messageTemplate, T0 property0, T1 property1)
-        => Write(LogLevel.Error, exception, messageTemplate, property0, property1);
+    public void Error<T0, T1>(Exception exception, string messageTemplate, T0 property0, T1 property1, bool writeToEventLog)
+        => Write(LogLevel.Error, exception, messageTemplate, property0, property1, writeToEventLog);
 
-    public void Error<T0, T1, T2>(Exception exception, string messageTemplate, T0 property0, T1 property1, T2 property2)
-        => Write(LogLevel.Error, exception, messageTemplate, property0, property1, property2);
+    public void Error<T0, T1, T2>(Exception exception, string messageTemplate, T0 property0, T1 property1, T2 property2, bool writeToEventLog)
+        => Write(LogLevel.Error, exception, messageTemplate, property0, property1, property2, writeToEventLog);
 
-    public void Error(Exception exception, string messageTemplate, object[] args)
-        => Write(LogLevel.Error, exception, messageTemplate, args);
+    public void Error(Exception exception, string messageTemplate, object[] args, bool writeToEventLog)
+        => Write(LogLevel.Error, exception, messageTemplate, args, writeToEventLog);
 
-    private void Write<T>(LogLevel level, Exception? exception, string messageTemplate, T property)
+    private static void WriteEventSourceLog(LogLevel level, string message)
+    {
+        switch (level)
+        {
+            case LogLevel.Information:
+                AutoInstrumentationEventSource.Log.Information(message);
+                break;
+            case LogLevel.Debug:
+                AutoInstrumentationEventSource.Log.Verbose(message);
+                break;
+            case LogLevel.Warning:
+                AutoInstrumentationEventSource.Log.Warning(message);
+                break;
+            case LogLevel.Error:
+                AutoInstrumentationEventSource.Log.Error(message);
+                break;
+        }
+    }
+
+    private void Write<T>(LogLevel level, Exception? exception, string messageTemplate, T property, bool writeToEventLog)
     {
         if (IsEnabled(level))
         {
             // Avoid boxing + array allocation if disabled
-            WriteImpl(level, exception, messageTemplate, new object?[] { property });
+            WriteImpl(level, exception, messageTemplate, new object?[] { property }, writeToEventLog);
         }
     }
 
-    private void Write<T0, T1>(LogLevel level, Exception? exception, string messageTemplate, T0 property0, T1 property1)
+    private void Write<T0, T1>(LogLevel level, Exception? exception, string messageTemplate, T0 property0, T1 property1, bool writeToEventLog)
     {
         if (IsEnabled(level))
         {
             // Avoid boxing + array allocation if disabled
-            WriteImpl(level, exception, messageTemplate, new object?[] { property0, property1 });
+            WriteImpl(level, exception, messageTemplate, new object?[] { property0, property1 }, writeToEventLog);
         }
     }
 
-    private void Write<T0, T1, T2>(LogLevel level, Exception? exception, string messageTemplate, T0 property0, T1 property1, T2 property2)
+    private void Write<T0, T1, T2>(LogLevel level, Exception? exception, string messageTemplate, T0 property0, T1 property1, T2 property2, bool writeToEventLog)
     {
         if (IsEnabled(level))
         {
             // Avoid boxing + array allocation if disabled
-            WriteImpl(level, exception, messageTemplate, new object?[] { property0, property1, property2 });
+            WriteImpl(level, exception, messageTemplate, new object?[] { property0, property1, property2 }, writeToEventLog);
         }
     }
 
-    private void Write(LogLevel level, Exception? exception, string messageTemplate, object[] args)
+    private void Write(LogLevel level, Exception? exception, string messageTemplate, object[] args, bool writeToEventLog)
     {
         if (IsEnabled(level))
         {
             // logging is not disabled
-            WriteImpl(level, exception, messageTemplate, args);
+            WriteImpl(level, exception, messageTemplate, args, writeToEventLog);
         }
     }
 
-    private void WriteImpl(LogLevel level, Exception? exception, string messageTemplate, object?[] args)
+    private void WriteImpl(LogLevel level, Exception? exception, string messageTemplate, object?[] args, bool writeToEventLog)
     {
         try
         {
-            var message =
-                $"[{DateTime.UtcNow:O}] [{level}] {string.Format(messageTemplate, args)} {Environment.NewLine}";
-            _sink.Write(message);
-
+            var rawMessage = string.Format(messageTemplate, args);
             if (exception != null)
             {
-                var exceptionMessage = $"Exception: {exception.Message}{Environment.NewLine}{exception}{Environment.NewLine}";
-                _sink.Write(exceptionMessage);
+                rawMessage += $"{Environment.NewLine}Exception: {exception.Message}{Environment.NewLine}{exception}";
+            }
+
+            _sink.Write($"[{DateTime.UtcNow:O}] [{level}] {rawMessage} {Environment.NewLine}");
+            if (writeToEventLog)
+            {
+                WriteEventSourceLog(level, rawMessage);
             }
         }
         catch